refactor(frontend): split SSOProvidersList god component (ATT-430)
Extract the ~1191-line SSOProvidersList into focused units while
preserving UI and behavior:

- formDefaults.ts: constants, default config factories, helpers
- useSSOProviderForm: form state, load effect, handlers, submit
- useSSOProviderSetupUrls: callback/regex/docs URL memos
- SSOProvidersTable: provider list table + empty state
- form/OIDCConfigForm, form/SAMLConfigForm: per-type config sections
- form/PermissionMappingsSection: shared permission-mapping inputs
- form/SetupInstructionsSection: setup URLs + copy actions
- SSOProviderFormDrawer: drawer assembling the form
- SSOProvidersList: thin orchestrator wiring hooks + subcomponents
